Resync pandas stubs from pylance
