+++ pandas/_libs/tslibs/timedeltas.pyi
@@ -50,5 +50,34 @@ class Timedelta(_Timedelta):
     def __divmod__(self, other) -> Tuple[int, int]: ...
     def __rdivmod__(self, other) -> Tuple[int, int]: ...
 
+    @property
+    def asm8(self) -> int: ...
+    @property
+    def components(self) -> int: ...
+    @property
+    def days(self) -> int: ...
+    @property
+    def delta(self) -> int: ...
+    @property
+    def microseconds(self) -> int: ...
+    @property
+    def nanoseconds(self) -> int: ...
+    @property
+    def resolution_string(self) -> str: ...
+    @property
+    def seconds(self) -> int: ...
+    max: Timedelta = ...
+    min: Timedelta = ...
+    resolution: 'Timedelta' = ...
+    def ceil(self, freq, **kwargs) -> Timedelta: ...
+    def floor(self, freq, **kwargs) -> Timedelta: ...
+    def isoformat(self) -> str: ...
+    def round(self, freq) -> Timedelta: ...
+    def to_numpy(self) -> _np.timedelta64: ...
+    def to_pytimedelta(self) -> datetime.timedelta: ...
+    def to_timedelta64(self) -> _np.timedelta64: ...
+    def total_seconds(self) -> int: ...
+
+
 def delta_to_nanoseconds(delta: _Timedelta) -> int: ...
 def ints_to_pytimedelta(arr: Sequence[int], box: bool = ...) -> np.ndarray: ...
