+++ pandas/core/tools/__init__.pyi
@@ -1,0 +1,3 @@
+from .datetimes import to_datetime as to_datetime
+from .timedeltas import to_timedelta as to_timedelta
+
